add accordion docs
