+++ src/docs/data/components/accordion-headless.ts
@@ -1,0 +1,15 @@
+import type { ExampleHighlights, Data } from "../types";
+
+const highlights: ExampleHighlights = {
+    main: {
+        'app.svelte': {
+            lines: '',
+            words: [{ strings: [''], id: 'builder' }],
+        }
+    }
+}
+
+
+export const accordionHeadlessData: Data = {
+    highlights,
+}
