+++ src/docs/guides/components/accordion-headless-old.md
@@ -1,6 +1,5 @@
 ---
-title: Accordion
-description: A headless accordion component.
+
 ---
 
 <script>
