+++ src/docs/previews/accordion/colors/app.svelte
@@ -6,6 +6,7 @@
         title: string;
         content: string;
         icon: string;
+        titleBgStyle: string;
     }
 
     // From the Avatar wiki: https://avatar.fandom.com/wiki/Bending_arts
@@ -14,32 +15,36 @@
             key: 'water',
             title: 'Waterbending',
             content: 'Waterbending is practiced by some people of the Water Tribe. A versatile element, it is unique in the sense that the original bender of the element was not an animal, but rather the Moon.',
-            icon: 'i-material-symbols-water-drop-outline-rounded'
+            icon: 'i-material-symbols-water-drop-outline-rounded',
+            titleBgStyle: 'transition-all primary-400 hover:primary-500'
         },
         {
             key: 'earth',
             title: 'Earthbending',
             content: 'Earthbending originates in the Earth Kingdom and the first earthbenders were badgermoles, who are said to love music.',
-            icon: 'i-mdi-earth'
+            icon: 'i-mdi-earth',
+            titleBgStyle: 'transition-all secondary-400 hover:secondary-500'
         },
         {
             key: 'fire',
             title: 'Firebending',
             content: 'Firebending is used by the people of the Fire Nation and is the most aggressive bending art. Dragons were the first firebenders; they subsequently taught the Sun Warriors.',
-            icon: 'i-material-symbols-local-fire-department-rounded'
+            icon: 'i-material-symbols-local-fire-department-rounded',
+            titleBgStyle: 'bg-gradient-br-primary-tertiary text-on-primary-500 opacity-80 hover:opacity-100'
         },
         {
             key: 'air',
             title: 'Airbending',
             content: 'Airbending is the bending art used by the Air Nomads; the flying bison were the original airbenders.',
-            icon: 'i-mdi-weather-windy-variant'
+            icon: 'i-mdi-weather-windy-variant',
+            titleBgStyle: 'border-1 border-surface-400/80 hover:(primary-500 border-primary-500)'
         }
     ];
 </script>
 
-<Accordion width="w-92">
-    {#each items as { title, content, key, icon }}
-        <AccordionItem {key} >
+<Accordion width="w-92" >
+    {#each items as { title, content, key, titleBgStyle, icon }}
+        <AccordionItem {key} {titleBgStyle} >
             <svelte:fragment slot="title">
                 <div class="flex justify-left items-center gap-2">
                     <span class="text-lg {icon}" />
