+++ src/lib/components/accordion/index.ts
@@ -1,0 +1,5 @@
+
+export { default as Accordion } from './accordion.svelte';
+export { default as AccordionItem } from './accordion-item.svelte';
+
+export * from './types';
