+++ src/lib/components/combobox/index.ts
@@ -1,4 +1,4 @@
 
-export { default as Combobox } from './combo-box.svelte';
+export { default as Combobox } from './combobox.svelte';
 
 export * from './types';
