+++ src/lib/components/index.ts
@@ -1,4 +1,5 @@
 
+export * from './accordion';
 export * from './combobox';
 export * from './dialog';
 // export * from './popover';
