+++ src/lib/index.ts
@@ -1,24 +1,12 @@
-// Reexport your entry components here
 
 // Types
 export type { DropMenuGroup } from './components/dropdown-menu/types';
-// export type { BackdropTransition } from './components/modal/types.d';
-// export type { RadioItem } from './components/radio-group/types.d';
-// export type { Heading, TOCIndentStyles, ToC, TOCType } from './components/table-of-contents/types.d';
-// export type { TabHeader } from './components/tabs/types';
 
 // Components
 export * from './components';
 
-export { default as Accordion } from './components/accordion/Accordion.svelte';
-export { default as AccordionItem } from './components/accordion/AccordionItem.svelte';
+// export { default as Accordion } from './components/accordion/Accordion.svelte';
+// export { default as AccordionItem } from './components/accordion/AccordionItem.svelte';
 export { default as Avatar } from './components/avatar/Avatar.svelte';
-// export { default as Combobox } from './components/combobox/Combobox.svelte';
 export { default as DropMenu } from './components/dropdown-menu/DropMenu.svelte';
-// export { default as Modal } from './components/modal/Modal.svelte';
-// export { default as RadioGroup } from './components/radio-group/RadioGroup.svelte';
 export { default as ToggleSwitch } from './components/toggle-switch/ToggleSwitch.svelte';
-
-// export { default as TableOfContents } from './components/table-of-contents/TableOfContents.svelte';
-// Headless
-// export { create_toc, default_indentation_styles, scroll_to_element } from './components/table-of-contents/toc';
